requirements.txt
feedparser
requests
beautifulsoup4
trafilatura
faiss-cpu
python-dotenv
google-genai
lxml
fastapi
uvicorn
lxml_html_clean
azure-search-documents
azure-core